pyproject.toml
[project]
name = "my-travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.22.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
